Codechange: Large sync from parallel work - genworld, industry, viewport, GPU pipeline
Merge extensive parallel changes: genworld cache removal, industry
optimizations, lake generation updates, viewport improvements, motion
vector hardening, OpenGL pipeline polish, screenshot system updates,
render backend refinements, and test updates.

+++ src/video/video_driver.hpp
@@ -70,7 +70,7 @@ extern uint8_t _video_pixel_smooth_amount;
 extern bool _video_auto_supersample;
 extern bool _video_fake_shadows;
 extern uint8_t _video_shadow_intensity;
-extern uint8_t _video_shadow_angle;
+extern uint16_t _video_shadow_angle;
 extern uint8_t _video_shadow_length;
 extern uint8_t _video_shadow_softness;
 extern bool _video_water_reflections;
